+++ client/src/io/worxbend/gitea4s/internal/SttpGiteaClient.scala
@@ -291,6 +291,9 @@ final class SttpGiteaClient(config: GiteaConfig, backend: Backend[Task]) extends
   override def languages(owner: String, repo: String): IO[GiteaError, LanguageStatistics] =
     executor.send(GiteaRequests.repoLanguages(config, owner, repo))
 
+  override def gpgSigningKey(owner: String, repo: String): IO[GiteaError, String] =
+    executor.send(GiteaRequests.repoSigningKey(config, owner, repo))
+
   override def tag(owner: String, repo: String, tag: String): IO[GiteaError, Tag] =
     executor.send(GiteaRequests.repoTag(config, owner, repo, tag))